+++ CHANGELOG.md
@@ -9,6 +9,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ### Added
 
+- Added the Norwegian summarisation datasets NorSumm-nb and NorSumm-nn, based on the
+  [NorSumm dataset](https://github.com/SamiaTouileb/NorSumm). The splits are given by
+  14 / 16 / 33 samples for train / val / test, respectively. Both datasets are marked as
+  `unofficial` for now.
 - Added the new Danish linguistic acceptability dataset DaLA. It's marked as 
   unofficial for now. This was added by @N-essuno ✨